+++ cpp/tests/test_solver.cpp
@@ -5,10 +5,10 @@
 #include <toppra/solver/glpk-wrapper.hpp>
 #endif
 
-#include <toppra/constraint/linear_joint_velocity.hpp>
 #include <toppra/constraint/linear_joint_acceleration.hpp>
-
+#include <toppra/constraint/linear_joint_velocity.hpp>
 #include <toppra/geometric_path.hpp>
+#include <toppra/geometric_path/piecewise_poly_path.hpp>
 
 #include "gtest/gtest.h"